test: split runtime boundary hotspots
