+++ tests/boundary/storage/recovery/test_enumeration_experiments.py
@@ -1,6 +1,9 @@
+"""Enumeration experiment lifecycle and contract boundary tests."""
+
 from __future__ import annotations
 
 import pytest
+from tests.boundary.storage.recovery.enumeration_experiments_support import _claim
 
 from jacobian.contracts.discovery import (
     EnumerationBudget,
@@ -8,112 +11,7 @@ from jacobian.contracts.discovery import (
     ExperimentState,
     SearchEnumerateRequest,
 )
-from jacobian.contracts.evaluation import EvaluationBatchResult, EvaluationProfile
-from jacobian.contracts.plugins import PluginManifest
-from jacobian.contracts.results import (
-    Execution,
-    ExecutionStatus,
-    InputStatus,
-    InputValidation,
-)
 from jacobian.experiments import ExperimentError, ExperimentNotFoundError
-from jacobian.runtime.model import JacobianRuntime
-from jacobian.storage.errors import StorageError
-
-
-def _claim(
-    runtime: JacobianRuntime,
-    *,
-    reference_name: str,
-    predicate: str,
-    parameters: dict[str, object],
-) -> tuple[str, str]:
-    reference = runtime.portfolio.references[reference_name]
-    claim = runtime.core.artifacts.put(
-        schema_uri=reference.claim_schema_uri,
-        semantics_uri=reference.semantics_uri,
-        payload={
-            "claim_schema_version": "1",
-            "domain_id": (
-                "jacobian.graph-paths"
-                if reference_name == "graph_paths"
-                else "jacobian.integer-matrices"
-            ),
-            "domain_version": "1",
-            "semantics_uri": reference.semantics_uri,
-            "quantifiers": [],
-            "predicate": {"name": predicate, "parameters": parameters},
-            "bounds": {},
-            "required_capabilities": ["CandidateEnumerator", "Evaluator"],
-            "correspondence_status": "HUMAN_REVIEWED",
-        },
-    )
-    return claim.artifact_uri, reference.plugin_id
-
-
-def _install_matrix_enumerator_plugin(
-    runtime: JacobianRuntime,
-    *,
-    entrypoint: str,
-    evaluator_entrypoint: str = "jacobian.plugins.matrices:evaluate_capability",
-) -> str:
-    matrix = runtime.portfolio.references["matrices"]
-    enumerator = runtime.core.plugins.register_implementation(entrypoint)
-    evaluator = runtime.core.plugins.register_implementation(evaluator_entrypoint)
-    manifest = runtime.core.artifacts.put(
-        schema_uri=runtime.services.reference_installer.manifest_schema_uri,
-        semantics_uri=runtime.services.reference_installer.manifest_semantics_uri,
-        payload=PluginManifest(
-            domain_id="jacobian.integer-matrices",
-            domain_version="1",
-            semantics_uri=matrix.semantics_uri,
-            claim_schema_uri=matrix.claim_schema_uri,
-            candidate_schema_uri=matrix.candidate_schema_uri,
-            capabilities={
-                "CandidateEnumerator": {
-                    "implementation_uri": enumerator,
-                    "entrypoint": entrypoint,
-                    "version": "1",
-                },
-                "Evaluator": {
-                    "implementation_uri": evaluator,
-                    "entrypoint": evaluator_entrypoint,
-                    "version": "1",
-                },
-            },
-        ).model_dump(mode="json"),
-    )
-    runtime.core.plugins.install(manifest.artifact_uri)
-    return manifest.artifact_uri
-
-
-def _matrix_claim_for_plugin(
-    runtime: JacobianRuntime,
-    *,
-    plugin_id: str,
-) -> str:
-    matrix = runtime.portfolio.references["matrices"]
-    claim = runtime.core.artifacts.put(
-        schema_uri=matrix.claim_schema_uri,
-        semantics_uri=matrix.semantics_uri,
-        payload={
-            "claim_schema_version": "1",
-            "domain_id": "jacobian.integer-matrices",
-            "domain_version": "1",
-            "semantics_uri": matrix.semantics_uri,
-            "quantifiers": [],
-            "predicate": {"name": "is_nonsingular", "parameters": {}},
-            "bounds": {},
-            "required_capabilities": ["CandidateEnumerator", "Evaluator"],
-            "correspondence_status": "HUMAN_REVIEWED",
-        },
-    )
-    validation = runtime.services.claims.validate(
-        claim_uri=claim.artifact_uri,
-        plugin_id=plugin_id,
-    )
-    assert validation.valid
-    return claim.artifact_uri
 
 
 def test_unknown_experiment_error_explains_recovery(
@@ -484,228 +382,3 @@ def test_cancelling_a_terminal_experiment_does_not_change_it(
     assert completed.state == ExperimentState.COMPLETED
     assert cancelled.accepted is False
     assert after == completed
-
-
-def test_enumerator_candidate_is_validated_before_archival(
-    authorized_complete_runtime,
-) -> None:
-    plugin_id = _install_matrix_enumerator_plugin(
-        authorized_complete_runtime,
-        entrypoint="tests.component.plugins._fixture_plugins:enumerate_invalid_candidate",
-    )
-    claim_uri = _matrix_claim_for_plugin(
-        authorized_complete_runtime,
-        plugin_id=plugin_id,
-    )
-
-    handle = authorized_complete_runtime.services.experiments.start_enumeration(
-        SearchEnumerateRequest(
-            claim_uri=claim_uri,
-            plugin_id=plugin_id,
-            bounds={"fixture": True},
-            budget=EnumerationBudget(
-                candidates_max=1,
-                wall_seconds=30,
-                page_size=1,
-            ),
-        )
-    )
-    snapshot = authorized_complete_runtime.services.experiments.wait(
-        handle.experiment_uri,
-        timeout_seconds=45,
-    )
-
-    assert snapshot.state == ExperimentState.ERROR
-    assert snapshot.stop_reason == EnumerationStopReason.ERROR
-    assert snapshot.coverage.value == "BOUNDED"
-    assert snapshot.verification.value == "UNVERIFIED"
-    assert snapshot.accounting.raw_candidates == 0
-    assert snapshot.archive_page_uris == ()
-
-
-def test_enumerator_timeout_remains_a_bounded_nonconclusion(
-    authorized_complete_runtime,
-) -> None:
-    plugin_id = _install_matrix_enumerator_plugin(
-        authorized_complete_runtime,
-        entrypoint="tests.component.plugins._fixture_plugins:wait_forever",
-    )
-    claim_uri = _matrix_claim_for_plugin(
-        authorized_complete_runtime,
-        plugin_id=plugin_id,
-    )
-    handle = authorized_complete_runtime.services.experiments.start_enumeration(
-        SearchEnumerateRequest(
-            claim_uri=claim_uri,
-            plugin_id=plugin_id,
-            bounds={"fixture": "timeout"},
-            budget=EnumerationBudget(
-                candidates_max=1,
-                wall_seconds=1,
-                page_size=1,
-            ),
-        )
-    )
-
-    with pytest.raises(
-        TimeoutError,
-        match="Inspect it or wait again with a larger timeout",
-    ):
-        authorized_complete_runtime.services.experiments.wait(
-            handle.experiment_uri, timeout_seconds=0
-        )
-
-    snapshot = authorized_complete_runtime.services.experiments.wait(
-        handle.experiment_uri,
-        timeout_seconds=15,
-    )
-
-    assert snapshot.state == ExperimentState.TIMEOUT
-    assert snapshot.stop_reason == EnumerationStopReason.WALL_TIME_LIMIT
-    assert snapshot.coverage.value == "BOUNDED"
-    assert snapshot.verification.value == "UNVERIFIED"
-    assert snapshot.accounting.raw_candidates == 0
-
-
-def test_evaluator_timeout_prevents_complete_enumeration_result(
-    authorized_complete_runtime,
-) -> None:
-    plugin_id = _install_matrix_enumerator_plugin(
-        authorized_complete_runtime,
-        entrypoint="jacobian.plugins.matrices:enumerate_candidates_capability",
-        evaluator_entrypoint="tests.component.plugins._fixture_plugins:wait_forever",
-    )
-    claim_uri = _matrix_claim_for_plugin(
-        authorized_complete_runtime,
-        plugin_id=plugin_id,
-    )
-    handle = authorized_complete_runtime.services.experiments.start_enumeration(
-        SearchEnumerateRequest(
-            claim_uri=claim_uri,
-            plugin_id=plugin_id,
-            bounds={"rows": 1, "cols": 1, "entries": [0]},
-            budget=EnumerationBudget(
-                candidates_max=1,
-                wall_seconds=1,
-                page_size=1,
-            ),
-        )
-    )
-
-    snapshot = authorized_complete_runtime.services.experiments.wait(
-        handle.experiment_uri,
-        timeout_seconds=15,
-    )
-
-    assert snapshot.state == ExperimentState.TIMEOUT
-    assert snapshot.stop_reason == EnumerationStopReason.WALL_TIME_LIMIT
-    assert snapshot.enumerator_reported_complete is False
-    assert snapshot.coverage.value == "BOUNDED"
-    assert snapshot.verification.value == "UNVERIFIED"
-
-
-def test_rejected_evaluation_batch_fails_enumeration(
-    authorized_complete_runtime,
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    claim_uri, plugin_id = _claim(
-        authorized_complete_runtime,
-        reference_name="matrices",
-        predicate="is_nonsingular",
-        parameters={},
-    )
-
-    def reject_batch(**_kwargs: object) -> EvaluationBatchResult:
-        return EvaluationBatchResult(
-            execution=Execution(status=ExecutionStatus.COMPLETED),
-            input=InputValidation(
-                status=InputStatus.REJECTED,
-                errors=("simulated incomplete evaluation",),
-            ),
-            claim_uri=claim_uri,
-            plugin_id=plugin_id,
-            profile=EvaluationProfile.EXACT_CANDIDATE,
-            seed=0,
-        )
-
-    monkeypatch.setattr(
-        authorized_complete_runtime.services.evaluation, "evaluate_batch", reject_batch
-    )
-    handle = authorized_complete_runtime.services.experiments.start_enumeration(
-        SearchEnumerateRequest(
-            claim_uri=claim_uri,
-            plugin_id=plugin_id,
-            bounds={"rows": 1, "cols": 1, "entries": [0]},
-            budget=EnumerationBudget(
-                candidates_max=1,
-                wall_seconds=30,
-                page_size=1,
-            ),
-        )
-    )
-
-    snapshot = authorized_complete_runtime.services.experiments.wait(
-        handle.experiment_uri, timeout_seconds=15
-    )
-
-    assert snapshot.state == ExperimentState.ERROR
-    assert snapshot.stop_reason == EnumerationStopReason.ERROR
-    assert snapshot.accounting.evaluated_candidates == 0
-    assert snapshot.archive_page_uris == ()
-    assert "artifact or plugin response was invalid" in snapshot.detail
-    assert "reference contract" in snapshot.detail
-    assert "simulated incomplete evaluation" not in snapshot.detail
-
-
-def test_terminal_archive_failure_marks_enumeration_error(
-    authorized_complete_runtime,
-    monkeypatch: pytest.MonkeyPatch,
-    caplog: pytest.LogCaptureFixture,
-) -> None:
-    claim_uri, plugin_id = _claim(
-        authorized_complete_runtime,
-        reference_name="matrices",
-        predicate="is_nonsingular",
-        parameters={},
-    )
-    original_put = (
-        authorized_complete_runtime.services.experiments._put_internal_artifact
-    )
-
-    def fail_terminal_archive(**kwargs: object) -> object:
-        if kwargs.get("summary") == "enumeration archive manifest":
-            raise StorageError("fixture archive failure")
-        return original_put(**kwargs)
-
-    monkeypatch.setattr(
-        authorized_complete_runtime.services.experiments,
-        "_put_internal_artifact",
-        fail_terminal_archive,
-    )
-    handle = authorized_complete_runtime.services.experiments.start_enumeration(
-        SearchEnumerateRequest(
-            claim_uri=claim_uri,
-            plugin_id=plugin_id,
-            bounds={"rows": 1, "cols": 1, "entries": [0]},
-            budget=EnumerationBudget(
-                candidates_max=1,
-                wall_seconds=30,
-                page_size=1,
-            ),
-        )
-    )
-
-    snapshot = authorized_complete_runtime.services.experiments.wait(
-        handle.experiment_uri,
-        timeout_seconds=15,
-    )
-
-    assert snapshot.state == ExperimentState.ERROR
-    assert snapshot.stop_reason == EnumerationStopReason.ERROR
-    assert snapshot.archive_uri is None
-    assert "could not save the final experiment archive" in snapshot.detail
-    assert "experiment remains unverified" in snapshot.detail
-    assert "StorageError" not in snapshot.detail
-    assert "runtime_ms" not in snapshot.detail
-    assert "fixture archive failure" not in snapshot.detail
-    assert "fixture archive failure" in caplog.text
